style(proxysql-agent): reorganized the project layout
- I followed some guides on go project layouts to redo the
  project layout
- Also enabled some extra golangci checks and addressed
  as many of the findings I could
