+++ .github/workflows/golangci-lint.yml
@@ -27,7 +27,7 @@ jobs:
           # Require: The version of golangci-lint to use.
           # When `install-mode` is `binary` (default) the value can be v1.2 or v1.2.3 or `latest` to use the latest version.
           # When `install-mode` is `goinstall` the value can be v1.2.3, `latest`, or the hash of a commit.
-          version: v1.54
+          version: v1.55.2
 
           # Optional: working directory, useful for monorepos
           # working-directory: somedir
