+++ configs/example_config.yaml
@@ -1,0 +1,52 @@
+# This application uses viper for configuration. Within viper there is a concept of overriding 
+# configuration. The order of precedence in these methods is as follows:
+#   1. Defaults set in the Configure() function
+#   2. Config file contens
+#   3. ENV variables
+#   4. Commandline flags
+#
+# When a comment says "defaults to X" that refers to the default set in Configure() (step 1 above)
+
+---
+# Set a pause of time of 0 seconds to allow the proxysql container to start; defaults to 1
+start_delay: 0
+
+log:
+  # Log level; follows log/slog conventions; defaults to INFO
+  level: "INFO"
+  # Log format; valid values are 'text' and 'JSON', defaults to JSON
+  format: "JSON"
+
+# ProxySQL admin connection configuration
+proxysql:
+  # Address for the proxysql admin interface; defaults to 127.0.0.1:6032
+  address: "127.0.0.1:6032"
+  # Username to connect to the admin interface; defaults to admin
+  username: "radmin"
+  # Password for the admin interface; no default set
+  password: "radmin"
+
+# The mode in which the agent should run, if any. valid values: [core OR satellite], no default
+# run_mode: core
+
+# Core mode specific configuration
+core:
+  # Number of seconds to pause in the loop; defaults to 10
+  interval: 10
+  # The k8s selector for the core pods. Currently does lookup based on a label, which is defined as:
+  #   spec:
+  #     template:
+  #       metadata:
+  #         labels:
+  #           app: proxysql
+  #           component: core  
+  podselector:
+    # Defaults to proxysql
+    app: proxysql
+    # Defaults to core
+    component: core
+
+# Satellite mode specific configuration
+satellite:
+  # The number of seconds to pause in the loop; defaults to 10
+  interval: 10
