--- example_config.yaml
@@ -1,34 +1,0 @@
----
-# set a pause of time of 5 seconds to allow the proxysql container to start; defaults to 1
-start_delay: 0
-
-# logging level for the app. follows log/slog conventions; defaults to INFO
-log_level: "INFO"
-
-# proxysql admin connection configuration
-proxysql:
-  # default: 127.0.0.1:6032
-  address: "127.0.0.1:6032"
-  # default: admin
-  username: "radmin"
-  # default: NONE
-  password: "radmin"
-
-# the mode in which the agent should run, if any. valid values: [core OR satellite], mutually exclusive, no default
-# run_mode: core
-
-# core mode specific configuratiobn
-core:
-  # number of seconds to pause in the loop. defaults to 10
-  interval: 10
-  # the k8s selector for the core pods
-  podselector:
-    # defaults to proxysql
-    app: proxysql
-    # defaults to core
-    component: core
-
-# satellite mode specific configuratiobn
-satellite:
-  # the number of seconds to pause in the loop. defaults to 10
-  interval: 10
