+++ internal/restapi/restapi_test.go
@@ -1,0 +1,1 @@
+package restapi
